+++ lib/datasafe/lib/model/target-alert-policy-association.ts
@@ -59,6 +59,10 @@ export interface TargetAlertPolicyAssociation {
    */
   "lifecycleState": model.AlertPolicyLifecycleState;
   /**
+   * Details about the current state of the target-alert policy association.
+   */
+  "lifecycleDetails"?: string;
+  /**
     * Free-form tags for this resource. Each tag is a simple key-value pair with no predefined name, type, or namespace. For more information, see [Resource Tags](https://docs.cloud.oracle.com/iaas/Content/General/Concepts/resourcetags.htm)
 * <p>
 Example: {@code {\"Department\": \"Finance\"}}